requirements.txt
Flask==2.3.3
pycryptodome==3.19.0
flask-limiter==3.5.0
APScheduler==3.10.4
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0